Точные количества и каскады задают контракт схемы
Остаток препарата и план лечения хранились в double precision. Половина таблетки
превращалась в 0.49999999999999994, ошибка накапливалась на каждом приёме, а тесты
пропорционального сжатия планов приходилось писать с допуском 0.001 — то есть
инвариант «сумма планов не превышает остаток» проверялся приблизительно.

Количества переведены на numeric(19,6) и BigDecimal. Округление — вниз: количество
здесь это обещание, что столько препарата действительно есть, и округление вверх
создавало бы остаток из воздуха. Нормализация живёт в сеттерах Drug и Using, поэтому
масштаб не расползается от умножений.

Пропорциональное сжатие планов теперь умножает до деления: отдельный коэффициент
60/90 пришлось бы округлить, и план 30 превращался бы в 19.999999 вместо 20. При
таком порядке точное частное получается там, где оно существует, а округление вниз
на каждом плане оставляет инвариант в силе.

Внешние ключи получили ON DELETE CASCADE со стороны препарата и аптечки. Со стороны
пользователя каскада намеренно нет: аптечки общие, и удаление пользователя молча
вынесло бы из чужой аптечки чужие планы. Определения ключей продублированы в
аннотациях, иначе схема Hibernate в тестах и db/schema.sql проверяли бы разные
каскады. Профиль dev переведён с update на validate: схему описывает schema.sql.

Проверки с допуском заменены на точные и проходят — арифметика стала точной не на
словах. Снапшот OpenAPI перегенерирован механически: из него ушли двенадцать
объявлений format: double.

+++ src/MedAppServer/src/main/kotlin/org/kert0n/medappserver/controller/DrugController.kt
@@ -21,6 +21,7 @@ import org.slf4j.LoggerFactory
 import org.springframework.http.HttpStatus
 import org.springframework.security.core.Authentication
 import org.springframework.web.bind.annotation.*
+import java.math.BigDecimal
 import java.util.*
 import io.swagger.v3.oas.annotations.parameters.RequestBody as SwaggerRequestBody
 
@@ -273,11 +274,11 @@ class DrugController(
 @Schema(description = "Drug quantity information")
 data class QuantityInfo(
     @Schema(description = "Actual quantity in stock")
-    val actualQuantity: Double,
+    val actualQuantity: BigDecimal,
     @Schema(description = "Total planned quantity across all treatment plans")
-    val plannedQuantity: Double,
+    val plannedQuantity: BigDecimal,
     @Schema(description = "Available quantity (actual - planned)")
-    val availableQuantity: Double
+    val availableQuantity: BigDecimal
 )
 
 @Schema(description = "Request to consume a drug")
@@ -285,7 +286,7 @@ data class ConsumeRequest(
     @NotNull
     @DecimalMin(value = "0.0", inclusive = false)
     @Schema(description = "Quantity to consume", example = "2.0", minimum = "0")
-    val quantity: Double
+    val quantity: BigDecimal
 )
 
 @Schema(description = "Request to move a drug to another medicine kit")
@@ -322,9 +323,9 @@ data class DrugDTO(
     @Schema(description = "Drug name", example = "Aspirin")
     val name: String,
     @Schema(description = "Current quantity", example = "100.0")
-    val quantity: Double,
+    val quantity: BigDecimal,
     @Schema(description = "Total planned quantity across treatment plans", example = "30.0")
-    val plannedQuantity: Double,
+    val plannedQuantity: BigDecimal,
     @Schema(description = "Quantity unit", example = "mg")
     val quantityUnit: String,
     @Schema(description = "Form type", example = "tablet")
@@ -351,7 +352,7 @@ data class DrugCreateDTO(
     @NotNull
     @DecimalMin("0.0")
     @Schema(description = "Quantity", example = "100.0", required = true, minimum = "0")
-    val quantity: Double,
+    val quantity: BigDecimal,
 
     @NotNull
     @Size(min = 1, max = 50)
@@ -391,7 +392,7 @@ data class DrugUpdateDTO(
 
     @DecimalMin("0.0")
     @Schema(description = "Quantity", example = "100.0", minimum = "0")
-    val quantity: Double? = null,
+    val quantity: BigDecimal? = null,
 
     @Size(min = 1, max = 50)
     @Schema(description = "Quantity unit", example = "mg")
